Fixed file structure and added student login
