+++ src/Services/teacher-auth-api-service.js
@@ -1,0 +1,49 @@
+import config from '../config'
+import TokenService from './token-service'
+
+const TeacherAuthApiService = {
+  postUser(user) {
+    return fetch(`${config.API_ENDPOINT}/user`, {
+      method: 'POST',
+      headers: {
+        'content-type': 'application/json',
+      },
+      body: JSON.stringify(user),
+    })
+      .then(res =>
+        (!res.ok)
+          ? res.json().then(e => Promise.reject(e))
+          : res.json()
+      )
+     
+  },
+  postLogin({ email, password }) {
+    return fetch(`${config.API_ENDPOINT}/auth/token`, {
+      method: 'POST',
+      headers: {
+        'content-type': 'application/json',
+      },
+      body: JSON.stringify({ email, password }),
+    })
+      .then(res =>
+        (!res.ok)
+          ? res.json().then(err => Promise.reject(err))
+          : res.json()
+      )
+  },
+  refreshToken() {
+    return fetch(`${config.API_ENDPOINT}/auth/token`, {
+      method: 'PUT',
+      headers: {
+        'authorization': `Bearer ${TokenService.getAuthToken()}`,
+      },
+    })
+      .then(res =>
+        (!res.ok)
+          ? res.json().then(e => Promise.reject(e))
+          : res.json()
+      )
+  },
+}
+
+export default TeacherAuthApiService
