--- src/services/auth-api-service.js
@@ -1,20 +1,0 @@
-import config from './config'
-
-const AuthApiService = {
-  postStudentLogin({ username }) {
-    return fetch(`${config.API_ENDPOINT}/auth/token`, {
-      method: 'POST',
-      headers: {
-        'content-type': 'application/json',
-      },
-      body: JSON.stringify({ username }),
-    })
-      .then(res =>
-        (!res.ok)
-          ? res.json().then(err => Promise.reject(err))
-          : res.json()
-      )
-  },
-}
-
-export default AuthApiService;
